requirements.txt
python-jobspy
pandas
python-dotenv
Flask
APScheduler
gunicorn
feedparser
pytz
